requirements.txt
boto3==1.9.86
botocore==1.12.86
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
dropbox==9.3.0
idna==2.8
jmespath==0.9.3
Pillow==5.4.1
python-dateutil==2.7.5
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
Wand==0.5.0
